refactored video component
